backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-allauth>=65.0,<66.0
django-cors-headers>=4.6,<5.0
channels>=4.2,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
celery>=5.4,<6.0
redis>=5.2,<6.0
psycopg[binary]>=3.2,<4.0
anthropic>=0.42,<1.0
python-dotenv>=1.0,<2.0
whitenoise
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-asyncio>=0.24,<1.0
factory-boy>=3.3,<4.0
channels[daphne]>=4.2,<5.0
